Switch back to pytorch 2.5.1
tempfile `ResourceWarning` is triggered by the implicit cleanup of
`_TEMP_DIR` created in `torch.distributed.nn.jit.instantiator.py`.
This warning is supressed by default but for some reason using `Join`
causes the warining to be displayed. I am not sure that it is a problem
though because afaik this implicit cleanup is the expected behavior.
Therefore it is safe to switch back to 2.5.1.